+++ R/clean_cache.R
@@ -3,6 +3,9 @@
 #' This function cleans the Conda cache by removing all packages and tarballs
 #' from the local cache directory. It is useful for freeing up disk space
 #' and ensuring that the cache does not contain outdated or unnecessary files.
+#' This function also removes any additional cache files created by the
+#' `condathis` package in the path reported by
+#' `tools::R_user_dir(package = "condathis", which = "cache")`.
 #'
 #' @param verbose A character string indicating the verbosity level of the
 #'  output. It can be one of "silent", "cmd", "output", or "full".
@@ -47,6 +50,16 @@ clean_cache <- function(
       )
     }
   )
+
+  r_user_cache <- fs::path(
+    tools::R_user_dir(package = "condathis", which = "cache")
+  )
+
+  if (fs::dir_exists(r_user_cache)) {
+    for (content_item in fs::dir_ls(r_user_cache)) {
+      fs::file_delete(content_item)
+    }
+  }
 
   if (isTRUE(verbose_list$strategy %in% c("full", "output"))) {
     cli::cli_inform(